+++ models/users/users.js
@@ -10,7 +10,9 @@ Model for Users Database
     var schema = new mongoose.Schema({
         username: { type: String, required: true, unique: true },
         password: { type: String, required: true },
-        status: { type: String, default: "user"} // Type of user [admin,User]
+        status: { type: String, default: "user"}, // Type of user [admin,User]
+        email: { type: String, required: true, unique: true },
+        activated: { type: Boolean, default: false }
     });
 
     var Users = mongoose.model ( "User", schema, "users2" );